+++ src/main/java/kitchenpos/order/domain/OrderService.java
@@ -3,7 +3,6 @@ package kitchenpos.order.domain;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.exception.NotFoundOrderException;
-import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.dto.OrderCreateRequest;
 import kitchenpos.order.dto.OrderResponse;
 import org.springframework.stereotype.Service;
@@ -11,16 +10,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class OrderService {
-    private final MenuRepository menuRepository;
     private final OrderRepository orderRepository;
     private final OrderValidator orderValidator;
 
 
-    public OrderService(
-            final MenuRepository menuRepository,
-            final OrderRepository orderRepository,
-            final OrderValidator orderValidator) {
-        this.menuRepository = menuRepository;
+    public OrderService(final OrderRepository orderRepository, final OrderValidator orderValidator) {
         this.orderRepository = orderRepository;
         this.orderValidator = orderValidator;
     }
@@ -28,20 +22,11 @@ public class OrderService {
     @Transactional
     public OrderResponse create(final OrderCreateRequest orderCreateRequest) {
         final Order order = orderCreateRequest.toOrder();
-        validateOrderLineItemMatchMenu(order);
 
         order.place(orderValidator);
 
         final Order saved = orderRepository.save(order);
         return OrderResponse.from(saved);
-    }
-
-    private void validateOrderLineItemMatchMenu(final Order order) {
-        final List<Long> menuIds = order.getOrderLineItems()
-                .stream()
-                .map(OrderLineItem::getMenuId)
-                .collect(Collectors.toList());
-        order.validateOrderLineItemSize(menuRepository.countByIdIn(menuIds));
     }
 
     @Transactional(readOnly = true)
