+++ src/main/java/kitchenpos/order/domain/OrderValidator.java
@@ -1,6 +1,9 @@
 package kitchenpos.order.domain;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import kitchenpos.exception.NotFoundOrderTableException;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.springframework.stereotype.Component;
@@ -9,12 +12,28 @@ import org.springframework.stereotype.Component;
 public class OrderValidator {
 
     private final OrderTableRepository orderTableRepository;
+    private final MenuRepository menuRepository;
 
-    public OrderValidator(final OrderTableRepository orderTableRepository) {
+    public OrderValidator(final OrderTableRepository orderTableRepository, final MenuRepository menuRepository) {
         this.orderTableRepository = orderTableRepository;
+        this.menuRepository = menuRepository;
     }
 
     public void validate(final Order order) {
+        validateOrderLineItemMatchMenu(order);
+        validateOrderTableIsOrderable(order);
+    }
+
+    private void validateOrderLineItemMatchMenu(final Order order) {
+        final List<Long> menuIds = order.getOrderLineItems()
+                .stream()
+                .map(OrderLineItem::getMenuId)
+                .collect(Collectors.toList());
+
+        order.validateOrderLineItemSize(menuRepository.countByIdIn(menuIds));
+    }
+
+    private void validateOrderTableIsOrderable(final Order order) {
         final OrderTable orderTable = orderTableRepository.findById(order.getOrderTableId())
                 .orElseThrow(NotFoundOrderTableException::new);
         orderTable.validateOrderable();
